# docs/generated/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Smart Delivery API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/shopify/callback": {
      "get": {
        "summary": "Complete Shopify OAuth callback",
        "operationId": "get_shopify_callback",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/shopify/install": {
      "get": {
        "summary": "Start Shopify OAuth install",
        "operationId": "get_shopify_install",
        "responses": {
          "302": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}